src/search/genomic_search.py
"""
Genomic search implementation for yak breeding research.
"""
import logging
from typing import Dict, List, Optional, Tuple

import pandas as pd
import numpy as np

logger = logging.getLogger(__name__)

class GenomicSearch:
    """Search functionality for genomic data."""

    def __init__(self):
        """Initialize GenomicSearch."""
        self.marker_index = {}
        self.region_index = {}

    def index_markers(self, markers: pd.DataFrame) -> None:
        """
        Index markers for efficient searching.

        Args:
            markers: DataFrame with marker information including IDs and positions
        """
        try:
            for _, row in markers.iterrows():
                self.marker_index[row['marker_id']] = {
                    'chromosome': row.get('chromosome'),
                    'position': row.get('position'),
                    'trait': row.get('trait'),
                    'effect': row.get('effect_size')
                }
            logger.info(f"Indexed {len(self.marker_index)} markers")
        except Exception as e:
            logger.error(f"Error indexing markers: {str(e)}")
            raise

    def search_by_marker(self, marker_id: str) -> Optional[Dict]:
        """
        Search for specific gene markers.

        Args:
            marker_id: ID of the marker to search for

        Returns:
            Dictionary with marker information if found, None otherwise
        """
        try:
            marker_info = self.marker_index.get(marker_id)
            if marker_info:
                logger.info(f"Found marker: {marker_id}")
                return marker_info
            logger.info(f"Marker not found: {marker_id}")
            return None
        except Exception as e:
            logger.error(f"Error searching for marker {marker_id}: {str(e)}")
            raise

    def search_by_region(self, chromosome: str,
                        start: int,
                        end: int) -> List[Dict]:
        """
        Search by chromosomal region.

        Args:
            chromosome: Chromosome identifier
            start: Start position
            end: End position

        Returns:
            List of markers in the specified region
        """
        try:
            region_markers = []
            for marker_id, info in self.marker_index.items():
                if (info['chromosome'] == chromosome and
                    start <= info['position'] <= end):
                    region_markers.append({
                        'marker_id': marker_id,
                        **info
                    })

            logger.info(f"Found {len(region_markers)} markers in region {chromosome}:{start}-{end}")
            return region_markers
        except Exception as e:
            logger.error(f"Error searching region {chromosome}:{start}-{end}: {str(e)}")
            raise

    def search_by_trait(self, trait: str,
                       min_effect: float = 0.0) -> List[Dict]:
        """
        Search for markers associated with a specific trait.

        Args:
            trait: Trait of interest
            min_effect: Minimum effect size threshold

        Returns:
            List of markers associated with the trait
        """
        try:
            trait_markers = []
            for marker_id, info in self.marker_index.items():
                if (info['trait'] == trait and
                    info['effect'] >= min_effect):
                    trait_markers.append({
                        'marker_id': marker_id,
                        **info
                    })

            logger.info(f"Found {len(trait_markers)} markers for trait {trait}")
            return trait_markers
        except Exception as e:
            logger.error(f"Error searching for trait {trait}: {str(e)}")
            raise
